# openapi/mini-commerce.v1.yaml
openapi: 3.1.0
info:
  title: Mini Commerce API
  version: 1.0.0
paths:
  /products:
    get:
      responses:
        '200': { description: Products, content: { application/json: { schema: { type: object, properties: { products: { type: array } } } } } }
        '503': { description: Database unavailable, content: { application/json: { schema: { type: object, properties: { error: { type: string } } } } } }
  /products/{id}/inventory:
    get:
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '200': { description: Inventory, content: { application/json: { schema: { type: object } } } }
        '404': { description: Product not found, content: { application/json: { schema: { type: object } } } }
        '503': { description: Database unavailable, content: { application/json: { schema: { type: object } } } }
  /orders:
    post:
      description: >-
        Internal caller API. Requires a unique Idempotency-Key per normalized order.
        Reusing a key with different items returns 409. The JSON body is limited to 32 KiB,
        100 item entries and a combined quantity of 100 per product; productId and quantity
        must be positive JSON integers. No customer identity or order ownership is enforced.
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, minLength: 1, maxLength: 128 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                items: { type: array }
      responses:
        '201': { description: Created order, content: { application/json: { schema: { type: object, properties: { order: { type: object } } } } } }
        '400': { description: Invalid order, content: { application/json: { schema: { type: object } } } }
        '413': { description: Request body exceeds 32 KiB }
        '415': { description: Unsupported charset or content encoding }
        '409': { description: Insufficient stock or conflicting idempotency key, content: { application/json: { schema: { type: object } } } }
        '503': { description: Database unavailable, content: { application/json: { schema: { type: object } } } }
  /orders/{id}:
    get:
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '200': { description: Order, content: { application/json: { schema: { type: object, properties: { order: { type: object, properties: { status: { type: string, enum: [CONFIRMED, CANCELLED] } } } } } } } }
        '404': { description: Order not found, content: { application/json: { schema: { type: object } } } }
        '503': { description: Database unavailable, content: { application/json: { schema: { type: object } } } }