# src/application.js
import express from 'express';
import { context, trace } from '@opentelemetry/api';
import { requestId, runWithRequestContext } from './request-context.js';

export function createApplication({ commerceService } = {}) {
  if (!commerceService) throw new TypeError('commerceService is required');
  const app = express();
  app.disable('x-powered-by');
  app.use((req, res, next) => {
    const spanContext = trace.getSpan(context.active())?.spanContext();
    const currentRequestId = requestId(req.get('x-request-id'));
    res.set('x-request-id', currentRequestId);
    runWithRequestContext({ requestId: currentRequestId, traceId: spanContext?.traceId }, next);
  });
  app.use(express.json({ limit: '32kb' }));
  app.get('/products', async (_req, res, next) => {
    try { res.json({ products: await commerceService.listProducts() }); } catch (error) { next(error); }
  });
  app.get('/products/:id/inventory', async (req, res, next) => {
    try { res.json(await commerceService.getInventory(req.params.id)); } catch (error) { next(error); }
  });
  app.post('/orders', async (req, res, next) => {
    try {
      const order = await commerceService.createOrder({
        idempotencyKey: req.get('Idempotency-Key') ?? '',
        items: req.body?.items,
      });
      res.status(201).json({ order });
    } catch (error) { next(error); }
  });
  app.get('/orders/:id', async (req, res, next) => {
    try { res.json({ order: await commerceService.getOrder(req.params.id) }); } catch (error) { next(error); }
  });
  app.use((_req, res) => res.status(404).json({ error: 'not found' }));
  app.use((error, _req, res, _next) => {
    if (res.headersSent) return _next(error);
    const parserErrors = {
      'entity.parse.failed': [400, 'invalid JSON body'],
      'entity.too.large': [413, 'request body too large'],
      'encoding.unsupported': [415, 'unsupported content encoding'],
      'charset.unsupported': [415, 'unsupported charset'],
      'request.aborted': [400, 'request aborted'],
      'request.size.invalid': [400, 'invalid request size'],
    };
    if (Object.hasOwn(parserErrors, error.type)) {
      const [status, message] = parserErrors[error.type];
      return res.status(status).json({ error: message });
    }
    const status = Number.isInteger(error.statusCode) && error.statusCode >= 400 && error.statusCode <= 599
      ? error.statusCode : 500;
    const message = error.name === 'DatabaseUnavailableError'
      ? 'database unavailable'
      : status >= 500 ? 'internal server error' : error.message;
    res.status(status).json({ error: message });
  });
  return app;
}